\name{microSTASIS}
\title{Microbiota STability ASsessment via Iterative cluStering}
\description{
The toolkit 'µSTASIS' has been developed for the stability analysis of 
microbiota in a temporal framework by leveraging on iterative clustering. 
Concretely, the core function uses Hartigan-Wong k-means algorithm as many times
as possible for stressing out paired samples from the same individuals to test 
if they remain together for multiple numbers of clusters over a whole data set 
of individuals. Moreover, the package includes multiple functions to subset 
samples from paired times, validate the results or visualize the output.
}