% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSheatmap.R
\name{mSheatmap}
\alias{mSheatmap}
\title{Plot a heatmap of the stability results.}
\usage{
mSheatmap(
  results,
  order = NULL,
  times,
  label = FALSE,
  low = "red2",
  mid = "yellow",
  high = "forestgreen",
  midpoint = 0.5
)
}
\arguments{
\item{results}{input data.frame resulting from \code{\link[=mSpreviz]{mSpreviz()}}.}

\item{order}{NULL object or character: none, mean or median; if the individuals should be sorted by any of those statistics of the stability values.}

\item{times}{character; names of the paired times to plot, i.e. colnames of results.}

\item{label}{logical; FALSE to avoid printing the mS score or TRUE to print it.}

\item{low}{color for the lowest value.}

\item{mid}{color for the middle value.}

\item{high}{color for the highest values.}

\item{midpoint}{value to situate the middle.}
}
\value{
A heatmap of the stability values in the form of a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot a heatmap of the stability results.
}
\examples{
times <- pairedTimes(data = clr, sequential = TRUE, common = "_0_")
mS <- iterativeClustering(pairedTimes = times, parallel = TRUE, common = "_")
results <- mSpreviz(results = mS, times = times)
mSheatmap(results = results, order = "mean", times = c("t1_t25", "t25_t26"), label = TRUE)
}
