% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpreviz.R
\name{mSpreviz}
\alias{mSpreviz}
\title{Process the [microSTASIS::iterativeClustering()] output to a new format ready for the implemented visualization functions.}
\usage{
mSpreviz(results, times)
}
\arguments{
\item{results}{list; output of [microSTASIS::iterativeClustering()].}

\item{times}{list; output of [microSTASIS::pairedTimes()].}
}
\value{
A data frame ready for its use under the implemented visualization functions and others.
}
\description{
Process the [microSTASIS::iterativeClustering()] output to a new format ready for the implemented visualization functions.
}
\examples{
times <- pairedTimes(data = clr, sequential = TRUE, common = "_0_")
mS <- iterativeClustering(pairedTimes = times, parallel = TRUE, common = "_0_")
results <- mSpreviz(results = mS, times = times)
}
