% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_times.R
\name{paired_times}
\alias{paired_times}
\title{Generate a subset matrix with paired times.}
\usage{
paired_times(data, first, second, common)
}
\arguments{
\item{data}{input matrix with each rowname including an ID, a common pattern and a sampling time.}

\item{first}{pattern associated with the first of the two sampling times.}

\item{second}{pattern associated with the second of the two sampling times.}

\item{common}{pattern that separates the ID and the sampling time.}
}
\value{
A matrix with the same number of columns as input and the samples from both samples times.
}
\description{
Generate a subset matrix with paired times.
}
\examples{
t1_t2 <- paired_times(data = clr, first = "_1",
                      second = "_25", common = "_0_")
}
