% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedTimes.R
\name{pairedTimes}
\alias{pairedTimes}
\title{Generate one or multiple matrices with paired times.}
\usage{
pairedTimes(data, ...)
}
\arguments{
\item{data}{input object: either a matrix with rownames including ID, common pattern and sampling time, or a TreeSummarizedExperiment object.}

\item{...}{The rest of possible arguments listed below.}

\item{sequential}{TRUE if paired times to analyse are sequential and present the desired alphanumerical order.}

\item{assay}{If class(data) == "TreeSummarizedExperiment", name of the assay to use.}

\item{alternativeExp}{If class(data) == "TreeSummarizedExperiment", name of the alternative experiment to use (if applicable).}

\item{ID}{If class(data) == "TreeSummarizedExperiment", one of the colData(data) colnames should be given as individuals.}

\item{timePoints}{If class(data) == "TreeSummarizedExperiment", one of the colData(data) colnames should be given as sampling times.}

\item{common}{If is.matrix(data), pattern that separates the ID and the sampling time in rownames.}

\item{specifiedTimePoints}{character vector to specify the selection of concrete paired times.}
}
\value{
A list of matrices with the same number of columns as input and with samples from paired sampling times as rows.
}
\description{
Generate one or multiple matrices with paired times.
}
\examples{
times <- pairedTimes(data = clr, sequential = TRUE, common = "_0_")
times_b <- pairedTimes(data = clr, sequential = FALSE, common = "_0_", 
                       specifiedTimePoints = c("1", "26"))

}
