% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clr}
\alias{clr}
\title{Detected ASV from multiple individuals at four different sampling times.}
\format{
A data frame with 131 rows and 226 variables
}
\usage{
clr
}
\description{
A dataset containing the amplicon sequence variants of 131 samples from the gut microbiota of 43 individuals. The values are transformed from counts by applying centred log-transformation (CLR).
}
\keyword{datasets}
