% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpreviz.R
\name{mSpreviz}
\alias{mSpreviz}
\title{Process the \code{\link[=iterativeClustering]{iterativeClustering()}} output to a new format
ready for the implemented visualization functions.}
\usage{
mSpreviz(results, times)
}
\arguments{
\item{results}{list; output of \code{\link[=iterativeClustering]{iterativeClustering()}}.}

\item{times}{list; output of \code{\link[=pairedTimes]{pairedTimes()}}.}
}
\value{
A data frame ready for its use under the implemented visualization
functions and others.
}
\description{
Process the \code{\link[=iterativeClustering]{iterativeClustering()}} output to a new format
ready for the implemented visualization functions.
}
\examples{
data(clr)
times <- pairedTimes(data = clr, sequential = TRUE, common = "_0_")
mS <- iterativeClustering(pairedTimes = times, common = "_")
results <- mSpreviz(results = mS, times = times)
}
